+++ app/components/packages/item/PackageItem.js
